feat(effects): move existing effects controller behind "Advanced" tab
